+++ Thesis/src/com/thesis/graph/Graph.java
@@ -3,15 +3,15 @@ package com.thesis.graph;
 import java.util.ArrayList;
 
 public class Graph {
-    private float[][] sparseM;
+    private double[][] sparseM;
     private ArrayList<SimpleNodeData> simpleNodeData;
 
-    public Graph(float[][] sparseM, ArrayList<SimpleNodeData> simpleNodeData) {
+    public Graph(double[][] sparseM, ArrayList<SimpleNodeData> simpleNodeData) {
         this.sparseM = sparseM;
         this.simpleNodeData = simpleNodeData;
     }
 
-    public float[][] getSparseM() {
+    public double[][] getSparseM() {
         return sparseM;
     }
